Fix total liquidity
